feat: add context support

+++ pkg/aws-s3/storage.go
@@ -2,6 +2,7 @@ package aws_s3
 
 import (
 	"bytes"
+	"context"
 	"fmt"
 	"io"
 	"time"
@@ -33,7 +34,11 @@ type AwsS3Storage struct {
 	config *AwsS3Config
 }
 
-func (s *AwsS3Storage) UploadFile(p goseidon.UploadFileParam) (*goseidon.UploadFileResult, error) {
+func (s *AwsS3Storage) UploadFile(ctx context.Context, p goseidon.UploadFileParam) (*goseidon.UploadFileResult, error) {
+	if ctx == nil {
+		return nil, fmt.Errorf("invalid context")
+	}
+
 	_, err := s.Client.PutObject(&s3.PutObjectInput{
 		Body:   bytes.NewReader(p.FileData),
 		Bucket: aws.String(s.config.BucketName),
@@ -48,7 +53,11 @@ func (s *AwsS3Storage) UploadFile(p goseidon.UploadFileParam) (*goseidon.UploadF
 	return res, nil
 }
 
-func (s *AwsS3Storage) RetrieveFile(p goseidon.RetrieveFileParam) (*goseidon.RetrieveFileResult, error) {
+func (s *AwsS3Storage) RetrieveFile(ctx context.Context, p goseidon.RetrieveFileParam) (*goseidon.RetrieveFileResult, error) {
+	if ctx == nil {
+		return nil, fmt.Errorf("invalid context")
+	}
+
 	out, err := s.Client.GetObject(&s3.GetObjectInput{
 		Key:    aws.String(p.Id),
 		Bucket: aws.String(s.config.BucketName),
@@ -68,7 +77,11 @@ func (s *AwsS3Storage) RetrieveFile(p goseidon.RetrieveFileParam) (*goseidon.Ret
 	return res, nil
 }
 
-func (s *AwsS3Storage) DeleteFile(p goseidon.DeleteFileParam) (*goseidon.DeleteFileResult, error) {
+func (s *AwsS3Storage) DeleteFile(ctx context.Context, p goseidon.DeleteFileParam) (*goseidon.DeleteFileResult, error) {
+	if ctx == nil {
+		return nil, fmt.Errorf("invalid context")
+	}
+
 	_, err := s.Client.DeleteObject(&s3.DeleteObjectInput{
 		Bucket: aws.String(s.config.BucketName),
 		Key:    aws.String(p.Id),
